DataRow:
Created the AllLotData object, which first scrapes the auction's soup, before it created all the LotData objects needed. Created a ScrapingBasedLotData object, which uses a soup to get its data. We do this in order not to have to keep instantiating the webdriver.

Need to make the lotRow and lotData system a lot simpler to understand; Not quite sure how yet.

Next step is to make the CTWebscraber object, so we can get a list of LIDs we need to use to webscrape. Maybe also instantiate the webdriver once so it stays instantiated? Otherwise, we might risk spending a lot of time closing and opening it.